Allow negative numbers in input.
Allow the user to specify that the input can display negative numbers
by showing a "-" in front of the number. By default, the component
does not allow negative numbers.

A number is detected as negative if it has an odd number of "-"
in it. Otherwise (no or an event number of "-"), it is detected
as positive.